Remove NumericIndexKey and use histogram point (#4743)
* Remove NumericIndexKey and use histogram point

* use old ordering

* fix ci

+++ lib/segment/src/index/field_index/numeric_index/mutable_numeric_index.rs
@@ -202,7 +202,7 @@ impl<T: Encodable + Numericable + Default> MutableNumericIndex<T> {
         let (decoded_idx, decoded_val) = T::decode_key(key);
         Point {
             val: decoded_val,
-            idx: decoded_idx as usize,
+            idx: decoded_idx,
         }
     }
